kernel: gdb: break features into individual files
Since gdbstub is able to do this, break individual features into their
own files.
